backend/app/requirements.txt
anyio==4.5.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.95.2
h11==0.14.0
httptools==0.6.4
idna==3.10
psycopg2-binary==2.9.5
pydantic==1.10.8
python-dotenv==1.0.0
PyYAML==6.0.2
requests==2.31.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.23.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-jose==3.3.0
email-validator
python-multipart
